+++ src/plugin.js
@@ -39,6 +39,29 @@ const plugin = {
         }
       ],
       since: "0.6.0"
+    },
+    xmlQuoteAttributes: {
+      type: "choice",
+      category: "XML",
+      default: "preserve",
+      description: "How to handle whitespaces in XML.",
+      choices: [
+        {
+          value: "preserve",
+          description:
+            "Quotes in attribute values will be preserved as written."
+        },
+        {
+          value: "single",
+          description:
+            "Quotes in attribute values will be converted to consistent single quotes and other quotes in the string will be escaped."
+        },
+        {
+          value: "double",
+          description:
+            "Quotes in attribute values will be converted to consistent double quotes and other quotes in the string will be escaped."
+        }
+      ]
     }
   },
   defaultOptions: {