+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- The `xmlQuoteAttributes: "preserve" | "single" | "double"` option has been added. See the README.
+
 ## [3.0.0] - 2023-07-06
 
 ### Changed